cmudict: separate out cmudict parsing
This makes the cmudict.parse function a wrapper around the dictionary
parser, with the original cmudict.parse moved to cmudict.parse_cmudict.
The warning to check mapping is now done in cmudict.parse.

This change is the first step in making the parser more flexible,
allowing it to support other formats using cmudict data such as the
festlex-cmu project.