cmudict: fix /AX/ phoneme validation
The /AX/ phoneme is the schwa (/@/) and so is never stressed (the
stressed version will either be /AH/ for the STRUT vowel or /ER/
for the NURSE vowel). As a result, it is common to treat this as a
consonant (excluding the stress marker).

Thus, the valid values for this phoneme are /AX/ and /AX0/.